Add the Logger Thread, not actually logging though
This thread is an alternative to external programs that we used in the
past, such as `script`. This is a pure-python implementation of this
such that we can get exactly the functionality that we need.

Due to the nature of logging a separate process, it makes a lot of sense
to have this running in a daemon / background thread. However, to make
this easier to program, the Logger Thread will not edit any data in the
main program thread, and will be the only thing that holds references to
it's own resources (e.g the log file handle).

The only external interaction that there should be beyond the starting
og the thread, is notifying it via `.stop()` that the process that it is
monitoring has stopped, although the logger thread may have noticed this
already.